Frontend fixes

+++ Entidades/U4Entrada.cs
@@ -9,8 +9,8 @@ namespace Entidades
     public sealed class U4Entrada
     {
         public string Funcion { get; set; }
-        public int PuntoA { get; set; }
-        public int PuntoB { get; set; }
+        public double PuntoA { get; set; }
+        public double PuntoB { get; set; }
         public int CantidadSubintervalos { get; set; }
         public int Metodo { get; set; }
     }
